Adds option for controling the basic-auth-behaviour;
